+++ app/public/utils/dataProcessing/dataProcessingUtils.js
@@ -41,3 +41,25 @@ export const rowDisplayStyle = (isMarked, shouldHideMarkedRecords) => isMarked
         ? '.none'
         : '.row-selected'
     : '.row-not-selected';
+
+export function getReadableFileSizeString(fileSizeInBytes) {
+    let i = -1;
+    const byteUnits = ['kB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
+    do {
+        fileSizeInBytes /= 1024;
+        i++;
+    } while (fileSizeInBytes > 1024);
+
+    return `${Math.max(fileSizeInBytes, 0.1).toFixed(1)}  ${byteUnits[i]}`;
+}
+
+/**
+ * Returns an object composed of the picked object properties
+ *
+ * @param {Object} obj The source object.
+ * @param {Array} keys The property paths to pick.
+ * @returns {Object} Returns the new object.
+ */
+export const pick = (obj, keys) => Object.fromEntries(Object.entries(obj)
+    .filter(([key]) => keys.includes(key)));
+// ToDo: add tests
